# small optimization to sre_compile._compile()