--- a/src/core/stats.h
+++ b/src/core/stats.h
@@ -28,9 +28,10 @@ namespace pyston {
 
 #define DISABLE_STATS 0
 
-#define STAT_ALLOCATIONS 0 && !DISABLE_STATS
-#define STAT_TIMERS 0 && !DISABLE_STATS
-#define EXPENSIVE_STAT_TIMERS 0 && STAT_TIMERS
+#define STAT_ALLOCATIONS (0 && !DISABLE_STATS)
+#define STAT_EXCEPTIONS (0 && !DISABLE_STATS)
+#define STAT_TIMERS (0 && !DISABLE_STATS)
+#define EXPENSIVE_STAT_TIMERS (0 && STAT_TIMERS)
 
 #if STAT_TIMERS
 #define STAT_TIMER(id, name)                                                                                           \