sre_compile.\_compile uses some constructs that are very expensive for us right now
(iteration over old-style classes without `__iter__`).  We should make that faster, but it
turns out that it was pretty easy+small+undetectable to just optimize sre_compile.py to not use it.

```
       django_template.py             8.1s (2)             8.1s (2)  -0.6%
            pyxl_bench.py             5.6s (2)             5.6s (2)  -0.6%
 sqlalchemy_imperative.py             2.7s (2)             2.7s (2)  -1.6%
        django_migrate.py             2.1s (2)             2.0s (2)  -3.2%
      virtualenv_bench.py             8.6s (2)             8.4s (2)  -2.1%
                  geomean                 4.7s                 4.6s  -1.6%
```